Move the Linux helpers into a separate file per Simon's PR feedback
